requirements.txt
inquirer
firebase-admin 
google-auth 
google-auth-oauthlib 
google-auth-httplib2
flask
python-dotenv